Move elementsd test to separate directory
