+++ Cargo.toml
@@ -29,10 +29,12 @@ actual-serde = { package = "serde", version = "1.0", optional = true }
 
 [dev-dependencies]
 serde_json = "1.0"
-elementsd = {version = "0.6.0", features=["0_21_0","bitcoind_22_0"]}
 actual-rand = { package = "rand", version = "0.8.4"}
-secp256k1 = {version = "0.24.0", features = ["rand-std"]}
 serde_test = "1.0.147"
+bitcoin = { version = "0.29.2", features = ["base64"] }
+secp256k1 = {version = "0.24.0", features = ["rand-std"]}
+actual-base64 = { package = "base64", version = "0.13.0" }
+
 
 [[example]]
 name = "htlc"
@@ -56,3 +58,4 @@ required-features = ["compiler"]
 
 [[example]]
 name = "psbt_sign_finalize"
+required-features = ["base64"]
