+++ bitcoind-tests/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "bitcoind-tests"
+version = "0.1.0"
+authors = ["sanket1729 <sanket1729@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+elements-miniscript = {path = "../"}
+elementsd = {version = "0.6.0", features=["0_21_0","bitcoind_22_0"]}
+actual-rand = { package = "rand", version = "0.8.4"}
+secp256k1 = {version = "0.24.0", features = ["rand-std"]}
