+++ bitcoind-tests/tests/test_csfs.rs
@@ -1,18 +1,19 @@
 //! # rust-miniscript integration test
 //!
-//! Arith expression fragment integration tests
+//! CheckSigFromStack integration tests
 //!
 
+use miniscript::{elements, bitcoin};
 use elements::pset::PartiallySignedTransaction as Psbt;
 use elements::sighash::SigHashCache;
 use elements::taproot::{LeafVersion, TapLeafHash};
 use elements::{
-    self, confidential, pset as psbt, secp256k1_zkp as secp256k1, sighash, OutPoint, Script,
+    confidential, pset as psbt, secp256k1_zkp as secp256k1, sighash, OutPoint, Script,
     Sequence, TxIn, TxOut, Txid,
 };
 use elementsd::ElementsD;
-use miniscript::psbt::{PsbtExt, PsbtInputExt};
-use miniscript::{Descriptor, ToPublicKey};
+use miniscript::psbt::{PsbtInputExt, PsbtInputSatisfier};
+use miniscript::{Descriptor, Satisfier, ToPublicKey};
 use rand::RngCore;
 mod setup;
 use setup::test_util::{self, TestData, PARAMS};
@@ -63,7 +64,7 @@ pub fn test_desc_satisfy(cl: &ElementsD, testdata: &TestData, desc: &str) -> Vec
         previous_output: outpoint,
         is_pegin: false,
         script_sig: Script::new(),
-        sequence: Sequence::from_consensus(1),
+        sequence: Sequence::from_height(1),
         asset_issuance: Default::default(),
         witness: Default::default(),
     };
@@ -102,7 +103,7 @@ pub fn test_desc_satisfy(cl: &ElementsD, testdata: &TestData, desc: &str) -> Vec
         Descriptor::TrExt(ref tr) => {
             let hash_ty = sighash::SchnorrSigHashType::Default;
 
-            let prevouts = [witness_utxo.clone()];
+            let prevouts = [witness_utxo];
             let prevouts = sighash::Prevouts::All(&prevouts);
             // ------------------ script spend -------------
             let x_only_keypairs_reqd: Vec<(secp256k1::KeyPair, TapLeafHash)> = tr
@@ -155,16 +156,37 @@ pub fn test_desc_satisfy(cl: &ElementsD, testdata: &TestData, desc: &str) -> Vec
     println!("Testing descriptor: {}", desc);
     // Finalize the transaction using psbt
     // Let miniscript do it's magic!
-    if let Err(e) = psbt.finalize_mall_mut(&secp, testdata.pubdata.genesis_hash) {
-        // All miniscripts should satisfy
-        panic!(
-            "Could not satisfy non-malleably: error{} desc:{} ",
-            e[0], desc
-        );
+    struct CsfsSatisfier<'a>(&'a TestData);
+
+    impl<'a> Satisfier<bitcoin::PublicKey> for CsfsSatisfier<'a> {
+        fn lookup_csfs_sig(
+            &self,
+            pk: &bitcoin::XOnlyPublicKey,
+            msg: &miniscript::extensions::CsfsMsg,
+        ) -> Option<secp256k1::schnorr::Signature> {
+            let xpk = pk.to_x_only_pubkey();
+            let known_xpks = &self.0.pubdata.x_only_pks;
+            let i = known_xpks.iter().position(|&x| x == xpk).unwrap();
+
+            // Create a signature
+            let keypair = &self.0.secretdata.x_only_keypairs[i];
+            let msg = secp256k1::Message::from_slice(&msg.as_inner()[..]).unwrap();
+            let mut aux_rand = [0u8; 32];
+            rand::thread_rng().fill_bytes(&mut aux_rand);
+
+            let secp = secp256k1::Secp256k1::new();
+            let sig = secp.sign_schnorr_with_aux_rand(&msg, keypair, &aux_rand);
+            Some(sig)
+        }
     }
-    let tx = psbt
-        .extract(&secp, testdata.pubdata.genesis_hash)
-        .expect("Extraction error");
+
+    let psbt_sat = PsbtInputSatisfier::new(&psbt, 0);
+    let csfs_sat = CsfsSatisfier(&testdata);
+
+    let mut tx = psbt.extract_tx().unwrap();
+    derived_desc
+        .satisfy(&mut tx.input[0], (psbt_sat, csfs_sat))
+        .expect("Satisfaction error");
 
     // Send the transactions to bitcoin node for mining.
     // Regtest mode has standardness checks
@@ -190,90 +212,20 @@ fn test_descs(cl: &ElementsD, testdata: &TestData) {
     // X!: X-only key with corresponding secret key unknown
 
     // Test 1: Simple spend with internal key
-    let wit = test_desc_satisfy(cl, testdata, "tr(X!,and_v(v:pk(X1),num64_eq(8,8)))");
+    let wit = test_desc_satisfy(cl, testdata, "tr(X!,csfs(X1,msg1))");
     assert!(wit.len() == 3);
 
-    let wit = test_desc_satisfy(cl, testdata, "tr(X!,and_v(v:pk(X1),num64_geq(9,8)))");
-    assert!(wit.len() == 3);
+    // test in a complicated miniscript
+    let wit = test_desc_satisfy(cl, testdata, "tr(X!,and_b(csfs(X2,msg2),a:csfs(X1,msg3)))");
+    assert!(wit.len() == 4);
 
-    let wit = test_desc_satisfy(cl, testdata, "tr(X!,and_v(v:pk(X1),num64_gt(9,8)))");
-    assert!(wit.len() == 3);
-
-    test_desc_satisfy(
-        cl,
-        testdata,
-        "tr(X!,and_v(v:pk(X1),num64_gt(9223372036854775807,9223372036854775806)))",
-    );
-
-    test_desc_satisfy(
-        cl,
-        testdata,
-        "tr(X!,and_v(v:pk(X1),num64_eq(inp_v(idx_sub(2,2)),100000000)))",
-    );
-    test_desc_satisfy(
-        cl,
-        testdata,
-        "tr(X!,and_v(v:pk(X1),num64_eq(inp_v(idx_sub(idx_div(4,idx_mul(2,2)),1)),100000000)))",
-    );
-    test_desc_satisfy(
-        cl,
-        testdata,
-        "tr(X!,and_v(v:pk(X1),num64_eq(inp_v(0),100000000)))",
-    );
-    test_desc_satisfy(
-        cl,
-        testdata,
-        "tr(X!,and_v(v:pk(X1),num64_eq(out_v(0),99997000)))",
-    );
-    test_desc_satisfy(
-        cl,
-        testdata,
-        "tr(X!,and_v(v:pk(X1),num64_eq(out_v(1),3000)))",
-    );
-    test_desc_satisfy(
-        cl,
-        testdata,
-        "tr(X!,and_v(v:pk(X1),num64_eq(inp_v(0),add(out_v(0),out_v(1)))))",
-    );
-    test_desc_satisfy(cl, testdata, "tr(X!,and_v(v:pk(X1),num64_leq(-10,-10)))");
-    test_desc_satisfy(cl, testdata, "tr(X!,and_v(v:pk(X1),num64_lt(-11,-10)))");
-
-    test_desc_satisfy(
-        cl,
-        testdata,
-        "tr(X!,and_v(v:pk(X1),num64_eq(add(6,2),mul(2,4))))",
-    );
-    test_desc_satisfy(
-        cl,
-        testdata,
-        "tr(X!,and_v(v:pk(X1),num64_eq(sub(3,3),div(0,9))))",
-    );
-    test_desc_satisfy(cl, testdata, "tr(X!,and_v(v:pk(X1),num64_eq(mod(9,3),0)))");
-    test_desc_satisfy(
-        cl,
-        testdata,
-        "tr(X!,and_v(v:pk(X1),num64_eq(bitand(0,134),0)))",
-    );
-    test_desc_satisfy(
-        cl,
-        testdata,
-        "tr(X!,and_v(v:pk(X1),num64_eq(bitor(1,3),3)))",
-    );
-    test_desc_satisfy(
-        cl,
-        testdata,
-        "tr(X!,and_v(v:pk(X1),num64_eq(bitxor(1,3),2)))",
-    );
-    test_desc_satisfy(
-        cl,
-        testdata,
-        "tr(X!,and_v(v:pk(X1),num64_eq(bitinv(0),-1)))",
-    );
-    test_desc_satisfy(cl, testdata, "tr(X!,and_v(v:pk(X1),num64_eq(neg(1),-1)))");
+    // test combining with other miniscript fragments
+    let wit = test_desc_satisfy(cl, testdata, "tr(X!,and_b(pk(X2),a:csfs(X1,msg4)))");
+    assert!(wit.len() == 4);
 }
 
 #[test]
-fn test_arith() {
+fn test_csfs() {
     let (cl, _, genesis_hash) = &setup::setup(false);
     let testdata = TestData::new_fixed_data(50, *genesis_hash);
     test_descs(cl, &testdata);
